Added Rook/Buy Menu
Buy menu needs a buildout - thats what Im working on next. the assets are in, things just need to be hooked up. I'm going to put in back buttons in and perhaps add a relic menu/some new graphics for it.
Rook moves as intended but isn't playing nice with the health bar system .