Level 2 pawn added | bomb ability added
Level 2 pawn is pretty simple right now.
Level 2 pawn is different in that is:
 - is larger
 - has more health
 - inflicts more damage
 - slowly regens
 - moves slower
 - selects actions more frequently

Bomb was manually added to the 1 key, not using the ability manager, though ideally I would be using it, I just don't know how to use it yet.
Bomb is also missing a good sprite, same with the explosion it triggers.